Extract context structures to dedicated context.go file
- Create context.go with execution context type definitions
- Move StepContext from step.go to context.go
- Move JobContext and SetFailed() method from job.go to context.go
- Improve separation of concerns and code organization
- Each file now has clearer responsibilities:
  - context.go: execution context data structures
  - job.go: job definition and execution logic
  - step.go: step definition and execution logic
